search components refactoring
