+++ src/components/ItemSearch/ItemSearch.tsx
@@ -1,24 +1,13 @@
-import { useCallback } from 'react';
-import { useSearchParams } from 'react-router-dom';
-import { useRecoilState, useSetRecoilState } from 'recoil';
+import { FC } from 'react';
 import { FormattedMessage } from 'react-intl';
-import state from '@state';
-import { getByIdentifier } from '@common/api/search.api';
-import { usePagination } from '@common/hooks/usePagination';
-import { normalizeLccn } from '@common/helpers/validations.helper';
-import { StatusType } from '@common/constants/status.constants';
-import { UserNotificationFactory } from '@common/services/userNotification';
-import { DEFAULT_PAGES_METADATA } from '@common/constants/api.constants';
 import { AdvancedSearchModal } from '@components/AdvancedSearchModal';
-import { DEFAULT_SEARCH_BY, SEARCH_RESULTS_LIMIT, SearchIdentifiers } from '@common/constants/search.constants';
+import { SEARCH_RESULTS_LIMIT } from '@common/constants/search.constants';
 import { DOM_ELEMENTS } from '@common/constants/domElementsIdentifiers.constants';
 import { SearchControls } from '@components/SearchControls';
 import { FullDisplay } from '@components/FullDisplay';
 import { Pagination } from '@components/Pagination';
-import { SearchResultList } from '@components/SearchResultList';
 import { SearchControlPane } from '@components/SearchControlPane';
-import { generateSearchParamsState } from '@common/helpers/search.helper';
-import { useLoadSearchResults } from '@common/hooks/useLoadSearchResults';
+import { useSearch } from '@common/hooks/useSearch';
 import GeneralSearch from '@src/assets/general-search.svg?react';
 import './ItemSearch.scss';
 
@@ -29,94 +18,45 @@ const EmptyPlaceholder = () => (
   </div>
 );
 
-export const ItemSearch = () => {
-  const setIsLoading = useSetRecoilState(state.loadingState.isLoading);
-  const [searchBy, setSearchBy] = useRecoilState(state.search.index);
-  const [query, setQuery] = useRecoilState(state.search.query);
-  const [message, setMessage] = useRecoilState(state.search.message);
-  const [data, setData] = useRecoilState(state.search.data);
-  const setStatusMessages = useSetRecoilState(state.status.commonMessages);
-  const setForceRefreshSearch = useSetRecoilState(state.search.forceRefresh);
+type ItemSearchProps = {
+  filters: SearchFilters;
+  hasSearchParams: boolean;
+  defaultSearchBy: SearchIdentifiers;
+  searchResultsListComponent: ReactElement;
+  isVisibleFullDisplay?: boolean;
+  isVisibleAdvancedSearch?: boolean;
+  isVisibleSearchByControl?: boolean;
+};
 
+export const ItemSearch: FC<ItemSearchProps> = ({
+  filters,
+  hasSearchParams,
+  defaultSearchBy,
+  searchResultsListComponent,
+  isVisibleFullDisplay = true,
+  isVisibleAdvancedSearch = true,
+  isVisibleSearchByControl = true,
+}) => {
   const {
-    getPageMetadata,
-    setPageMetadata,
-    getCurrentPageNumber,
-    setCurrentPageNumber,
+    submitSearch,
+    clearValues,
     onPrevPageClick,
     onNextPageClick,
-  } = usePagination(DEFAULT_PAGES_METADATA);
-  const currentPageNumber = getCurrentPageNumber();
-  const pageMetadata = getPageMetadata();
-  const setSearchParams = useSearchParams()?.[1];
-
-  const clearPagination = () => {
-    setPageMetadata(DEFAULT_PAGES_METADATA);
-    setCurrentPageNumber(0);
-  };
-
-  const clearMessage = useCallback(() => message && setMessage(''), [message]);
-
-  const validateAndNormalizeQuery = (type: SearchIdentifiers, query: string) => {
-    if (type === SearchIdentifiers.LCCN) {
-      const normalized = normalizeLccn(query);
-
-      !normalized && setMessage('marva.searchInvalidLccn');
-
-      return normalized;
-    }
-
-    return query;
-  };
-
-  const fetchData = async (query: string, searchBy: SearchIdentifiers, offset?: number) => {
-    clearMessage();
-
-    data && setData(null);
-
-    const updatedQuery = validateAndNormalizeQuery(searchBy, query);
-
-    if (!updatedQuery) return;
-
-    setIsLoading(true);
-
-    try {
-      const result = await getByIdentifier({ searchBy, query: updatedQuery as string, offset: offset?.toString() });
-      const { content, totalPages, totalRecords } = result;
-
-      if (!content.length) return setMessage('marva.searchNoRdsMatch');
-
-      setData(content);
-      setPageMetadata({ totalPages, totalElements: totalRecords });
-    } catch {
-      setStatusMessages(currentStatus => [
-        ...currentStatus,
-        UserNotificationFactory.createMessage(StatusType.error, 'marva.errorFetching'),
-      ]);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  useLoadSearchResults(fetchData);
-
-  const submitSearch = () => {
-    clearPagination();
-    setSearchParams(generateSearchParamsState(query, searchBy) as unknown as URLSearchParams);
-    setForceRefreshSearch(true);
-  };
-
-  const clearValues = () => {
-    clearPagination();
-    setData(null);
-    setSearchBy(DEFAULT_SEARCH_BY);
-    setQuery('');
-    setMessage('');
-  };
+    currentPageNumber,
+    pageMetadata,
+    message,
+    data,
+  } = useSearch({ hasSearchParams, defaultSearchBy });
 
   return (
     <div data-testid="id-search" className="item-search">
-      <SearchControls submitSearch={submitSearch} clearValues={clearValues} />
+      <SearchControls
+        submitSearch={submitSearch}
+        clearValues={clearValues}
+        filters={filters}
+        isVisibleSearchBy={isVisibleSearchByControl}
+        isVisibleAdvancedSearch={isVisibleAdvancedSearch}
+      />
       <div className={DOM_ELEMENTS.classNames.itemSearchContent}>
         <SearchControlPane />
         <div className={DOM_ELEMENTS.classNames.itemSearchContentContainer}>
@@ -127,7 +67,7 @@ export const ItemSearch = () => {
           )}
           {data && (
             <>
-              <SearchResultList />
+              {searchResultsListComponent}
               {pageMetadata.totalElements > 0 && (
                 <Pagination
                   currentPage={currentPageNumber}
@@ -143,8 +83,8 @@ export const ItemSearch = () => {
           {!data && !message && <EmptyPlaceholder />}
         </div>
       </div>
-      <FullDisplay />
-      <AdvancedSearchModal clearValues={clearValues} />
+      {isVisibleFullDisplay && <FullDisplay />}
+      {isVisibleAdvancedSearch && <AdvancedSearchModal clearValues={clearValues} />}
     </div>
   );
 };
