+++ src/components/SearchControls/SearchControls.tsx
@@ -2,15 +2,7 @@ import { ChangeEvent, FC, useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { useRecoilState, useResetRecoilState, useSetRecoilState } from 'recoil';
 import { FormattedMessage } from 'react-intl';
-import {
-  FiltersGroupCheckType,
-  FiltersType,
-  Format,
-  PublishDate,
-  SearchIdentifiers,
-  SearchLimiterNames,
-  Suppressed,
-} from '@common/constants/search.constants';
+import { SearchIdentifiers } from '@common/constants/search.constants';
 import { SEARCH_FILTERS_ENABLED } from '@common/constants/feature.constants';
 import { SearchQueryParams } from '@common/constants/routes.constants';
 import { Button, ButtonType } from '@components/Button';
@@ -22,81 +14,23 @@ import CaretDown from '@src/assets/caret-down.svg?react';
 import XInCircle from '@src/assets/x-in-circle.svg?react';
 import './SearchControls.scss';
 
-const filters = [
-  {
-    labelId: 'marva.publishDate',
-    type: FiltersGroupCheckType.Single,
-    children: [
-      {
-        id: PublishDate.AllTime,
-        type: FiltersType.Radio,
-        name: SearchLimiterNames.PublishDate,
-        labelId: 'marva.allTime',
-      },
-      {
-        id: PublishDate.TwelveMonths,
-        type: FiltersType.Radio,
-        name: SearchLimiterNames.PublishDate,
-        labelId: 'marva.past12Months',
-      },
-      {
-        id: PublishDate.FiveYears,
-        type: FiltersType.Radio,
-        name: SearchLimiterNames.PublishDate,
-        labelId: 'marva.past5Yrs',
-      },
-      {
-        id: PublishDate.TenYears,
-        type: FiltersType.Radio,
-        name: SearchLimiterNames.PublishDate,
-        labelId: 'marva.past10Yrs',
-      },
-    ],
-  },
-  {
-    labelId: 'marva.format',
-    type: FiltersGroupCheckType.Multi,
-    children: [
-      {
-        id: Format.Volume,
-        type: FiltersType.Checkbox,
-        name: SearchLimiterNames.Format,
-        labelId: 'marva.volume',
-      },
-      {
-        id: Format.Ebook,
-        type: FiltersType.Checkbox,
-        name: SearchLimiterNames.Format,
-        labelId: 'marva.onlineResource',
-      },
-    ],
-  },
-  {
-    labelId: 'marva.suppressed',
-    type: FiltersGroupCheckType.Single,
-    children: [
-      {
-        id: Suppressed.All,
-        type: FiltersType.Radio,
-        name: SearchLimiterNames.Suppressed,
-        labelId: 'marva.volume',
-      },
-      {
-        id: Suppressed.NotSuppressed,
-        type: FiltersType.Radio,
-        name: SearchLimiterNames.Suppressed,
-        labelId: 'marva.suppressed',
-      },
-    ],
-  },
-];
-
 type Props = {
   submitSearch: VoidFunction;
   clearValues: VoidFunction;
+  filters: SearchFilters;
+  isVisibleSearchBy?: boolean;
+  isVisibleAdvancedSearch?: boolean;
+  hasSearchParams?: boolean;
 };
 
-export const SearchControls: FC<Props> = ({ submitSearch, clearValues }) => {
+export const SearchControls: FC<Props> = ({
+  submitSearch,
+  clearValues,
+  filters,
+  isVisibleSearchBy = true,
+  isVisibleAdvancedSearch = true,
+  hasSearchParams = true,
+}) => {
   const [searchBy, setSearchBy] = useRecoilState(state.search.index);
   const [query, setQuery] = useRecoilState(state.search.query);
   const setMessage = useSetRecoilState(state.search.message);
@@ -119,7 +53,7 @@ export const SearchControls: FC<Props> = ({ submitSearch, clearValues }) => {
 
   const onResetButtonClick = () => {
     clearValuesAndResetControls();
-    setSearchParams({});
+    hasSearchParams && setSearchParams({});
     setNavigationState({});
   };
 
@@ -134,14 +68,16 @@ export const SearchControls: FC<Props> = ({ submitSearch, clearValues }) => {
         <CaretDown className="header-caret" />
       </div>
       <div className="inputs">
-        <Select
-          withIntl
-          id="id-search-select"
-          className="select-input"
-          value={searchBy}
-          options={Object.values(SearchIdentifiers)}
-          onChange={({ value }) => setSearchBy(value as SearchIdentifiers)}
-        />
+        {isVisibleSearchBy && (
+          <Select
+            withIntl
+            id="id-search-select"
+            className="select-input"
+            value={searchBy}
+            options={Object.values(SearchIdentifiers)}
+            onChange={({ value }) => setSearchBy(value as SearchIdentifiers)}
+          />
+        )}
         <Input
           id="id-search-input"
           type="text"
@@ -172,13 +108,15 @@ export const SearchControls: FC<Props> = ({ submitSearch, clearValues }) => {
         >
           <FormattedMessage id="marva.reset" />
         </Button>
-        <Button
-          type={ButtonType.Link}
-          className="search-button"
-          onClick={() => setIsAdvancedSearchOpen(isOpen => !isOpen)}
-        >
-          <FormattedMessage id="marva.advanced" />
-        </Button>
+        {isVisibleAdvancedSearch && (
+          <Button
+            type={ButtonType.Link}
+            className="search-button"
+            onClick={() => setIsAdvancedSearchOpen(isOpen => !isOpen)}
+          >
+            <FormattedMessage id="marva.advanced" />
+          </Button>
+        )}
       </div>
 
       {SEARCH_FILTERS_ENABLED && <SearchFilters filters={filters} />}
